backend/requirements.txt
# Core dependencies - pinned versions (frozen 2026-01-21)

# Vector search / LlamaIndex
llama-index-core==0.14.12
llama-index-embeddings-huggingface==0.6.1
llama-index-readers-file==0.5.6
llama-index-vector-stores-pinecone==0.7.1
pinecone==7.3.0

# NLP - spaCy
spacy==3.8.11
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

# Similarity metrics
Levenshtein==0.27.3
scikit-learn==1.8.0

# API Server
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21

# Text extraction
pypdf==6.6.0
python-docx==1.2.0
newspaper3k==0.2.8
lxml_html_clean==0.4.3  # Required by newspaper3k (split from lxml.html.clean)

# GitHub cloning
gitpython==3.1.46

# Authentication
pymongo==4.16.0
motor==3.7.1
beanie==2.0.1
python-jose[cryptography]==3.5.0
bcrypt==5.0.0

# Discord Bot
discord.py==2.6.4
nest-asyncio==1.6.0

# System metrics / HTTP
psutil==7.2.1
httpx==0.28.1

# ML models (transitive but pinned for stability)
sentence-transformers==5.2.0
transformers==4.57.6
torch==2.9.1

# NOTE: Streamlit runs separately (Streamlit Community Cloud)
